twitter_api/requirements.txt
django==1.11
MySQL-python==1.2.5
djangorestframework==3.6.2